refactor: fixed dataset details endpoint packaging
