+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/datasets/DatasetSearchResult.scala
@@ -28,6 +28,7 @@ import io.renku.graph.model.datasets.{Date, DatePublished, Description, Identifi
 import io.renku.graph.model.{GitLabUrl, projects}
 import io.renku.http.rest.Links.{Href, Link, Rel, _links}
 import io.renku.json.JsonOps._
+import io.renku.knowledgegraph.datasets.details.DatasetEndpoint
 import io.renku.tinytypes.constraints.NonNegativeInt
 import io.renku.tinytypes.json.TinyTypeEncoders._
 import io.renku.tinytypes.{IntTinyType, TinyTypeFactory}
