+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/datasets/details/BaseDetailsFinder.scala
@@ -17,6 +17,7 @@
  */
 
 package io.renku.knowledgegraph.datasets
+package details
 
 import cats.MonadThrow
 import cats.effect.Async
