+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/datasets/details/DatasetEndpointDocs.scala
@@ -17,6 +17,7 @@
  */
 
 package io.renku.knowledgegraph.datasets
+package details
 
 import Dataset._
 import cats.MonadThrow
