+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/datasets/details/DatasetFinder.scala
@@ -17,6 +17,7 @@
  */
 
 package io.renku.knowledgegraph.datasets
+package details
 
 import Dataset._
 import cats.MonadThrow
