+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/datasets/details/PartsFinder.scala
@@ -17,6 +17,7 @@
  */
 
 package io.renku.knowledgegraph.datasets
+package details
 
 import Dataset.DatasetPart
 import cats.MonadThrow
