+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/datasets/details/ProjectsFinder.scala
@@ -16,9 +16,8 @@
  * limitations under the License.
  */
 
-package io.renku.knowledgegraph.datasets
+package io.renku.knowledgegraph.datasets.details
 
-import Dataset.DatasetProject
 import cats.MonadThrow
 import cats.effect.kernel.Async
 import cats.syntax.all._
@@ -31,6 +30,7 @@ import io.renku.graph.model.datasets.Identifier
 import io.renku.graph.model.projects
 import io.renku.graph.model.projects.{Path, ResourceId, Visibility}
 import io.renku.http.server.security.model.AuthUser
+import io.renku.knowledgegraph.datasets.Dataset.DatasetProject
 import io.renku.triplesstore.SparqlQuery.Prefixes
 import io.renku.triplesstore._
 import org.typelevel.log4cats.Logger
