+++ knowledge-graph/src/test/scala/io/renku/knowledgegraph/datasets/details/DatasetEndpointDocsSpec.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.renku.knowledgegraph.datasets
+package io.renku.knowledgegraph.datasets.details
 
 import io.renku.config.renku
 import io.renku.generators.CommonGraphGenerators.renkuApiUrls
