+++ knowledge-graph/src/test/scala/io/renku/knowledgegraph/datasets/details/DatasetFinderSpec.scala
@@ -17,6 +17,7 @@
  */
 
 package io.renku.knowledgegraph.datasets
+package details
 
 import cats.effect.IO
 import cats.syntax.all._
